requirements.txt
behave==1.2.5
behave-django==0.3.0
Django==1.9.4
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
LinkHeader==0.4.3
parse==1.6.6
parse-type==0.3.4
six==1.10.0
uritemplate.py==0.3.0
